logging and replace all .cuda() with to(device) for cpu/gpu switch
